refactor: move ImageBuilderSubstitutor to its own class
